+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated jsonoid-discovery dependency to v0.7.3
 - Allow undefined properties to be required
+- Greatly simplify supported regexes for string patterns
 
 ### Fixed
 - Fix assembly task